+++ app/src/main/java/com/pict/pbl/medswift/screens/AppWidgets.kt
@@ -1,14 +1,15 @@
 package com.pict.pbl.medswift.ui.theme
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.slideInVertically
+import android.app.DatePickerDialog
+import android.widget.DatePicker
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.CalendarToday
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
@@ -16,14 +17,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import java.util.*
 
 @Composable
 fun ScreenTitle( title : String , icon : ImageVector? = null ) {
@@ -33,8 +35,7 @@ fun ScreenTitle( title : String , icon : ImageVector? = null ) {
             Text(
                 text = title ,
                 color = Color.Black ,
-                fontWeight = FontWeight.Bold ,
-                fontSize = 22.sp ,
+                style = MaterialTheme.typography.displayLarge ,
                 modifier = Modifier
                     .padding(24.dp)
                     .fillMaxWidth()
@@ -45,14 +46,12 @@ fun ScreenTitle( title : String , icon : ImageVector? = null ) {
         Text(
             text = title ,
             color = Color.Black ,
-            fontWeight = FontWeight.Bold ,
-            fontSize = 22.sp ,
+            style = MaterialTheme.typography.displayLarge ,
             modifier = Modifier
                 .padding(24.dp)
                 .fillMaxWidth()
         )
     }
-
     Divider(
         color = MaterialTheme.colorScheme.primary,
         modifier = Modifier
@@ -67,8 +66,7 @@ fun ScreenTitleWithoutDivider( title : String ) {
     Text(
         text = title ,
         color = Color.Black ,
-        fontWeight = FontWeight.Bold ,
-        fontSize = 22.sp ,
+        style = MaterialTheme.typography.titleLarge ,
         modifier = Modifier
             .padding(24.dp)
             .fillMaxWidth()
@@ -94,7 +92,7 @@ fun TextInput(
             onValueChange( it )
             inputString = it
         },
-        textStyle = TextStyle.Default.copy( fontSize = 20.sp ) ,
+        textStyle = MaterialTheme.typography.labelLarge ,
         colors = TextFieldDefaults.textFieldColors(
             focusedLabelColor = MaterialTheme.colorScheme.primary,
             unfocusedLabelColor = MaterialTheme.colorScheme.primary,
@@ -104,7 +102,7 @@ fun TextInput(
             unfocusedIndicatorColor = Color.Transparent,
             disabledIndicatorColor = Color.Transparent
         ),
-        label = { Text(text = label ) },
+        label = { Text(text = label , style = MaterialTheme.typography.labelSmall ) },
         singleLine = true,
         leadingIcon = if( icon != null ) { {Icon(imageVector = icon, contentDescription = label)} } else null ,
         keyboardOptions = KeyboardOptions(keyboardType = keyboardType, imeAction = ImeAction.Next),
@@ -112,21 +110,119 @@ fun TextInput(
     )
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DropDownInput(
+    modifier : Modifier = Modifier,
+    items : List<String> ,
+    onItemClick : ( (Int) -> Unit ) ,
+    label : String ,
+) {
+    var expanded by remember{ mutableStateOf(false) }
+    var selectedText by remember{ mutableStateOf( items[0]) }
+    Column( modifier = modifier ){
+        ExposedDropdownMenuBox(
+            expanded = expanded,
+            onExpandedChange = { expanded = !expanded },
+            modifier = Modifier.background(MaterialTheme.colorScheme.secondaryContainer , RoundedCornerShape(16.dp)),
+        ) {
+            TextField(
+                value = selectedText,
+                onValueChange = {},
+                readOnly = true,
+                textStyle = TextStyle.Default.copy( fontSize = 20.sp ) ,
+                colors = TextFieldDefaults.textFieldColors(
+                    focusedLabelColor = MaterialTheme.colorScheme.primary,
+                    unfocusedLabelColor = MaterialTheme.colorScheme.primary,
+                    textColor = Color.Black ,
+                    containerColor = MaterialTheme.colorScheme.secondaryContainer ,
+                    focusedIndicatorColor = Color.Transparent,
+                    unfocusedIndicatorColor = Color.Transparent,
+                    disabledIndicatorColor = Color.Transparent
+                ),
+                label = { Text(text = label) },
+                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) },
+                modifier = Modifier
+                    .padding(16.dp)
+                    .fillMaxWidth()
+                    .menuAnchor()
+            )
+            ExposedDropdownMenu(
+                expanded = expanded,
+                onDismissRequest = { expanded = false }
+            ) {
+                items.forEach{ item ->
+                    DropdownMenuItem(
+                        text = { Text(text = item , style = MaterialTheme.typography.labelSmall) },
+                        onClick = {
+                            selectedText = item
+                            expanded = false
+                            onItemClick( items.indexOf( item ) )
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
+
 
 @Composable
 fun BubbleText( text : String ,
                 modifier: Modifier = Modifier ,
-                fontSize : Int = 18 ,
-                textStyle: TextStyle = TextStyle( fontWeight = FontWeight.Bold ) ) {
+                fontSize : Int = 18  ) {
     Box( modifier = Modifier
         .padding(8.dp)
         .background(MaterialTheme.colorScheme.primaryContainer, RoundedCornerShape(16.dp)) )  {
         Text(
             text = text ,
             modifier = modifier ,
-            style = textStyle ,
             textAlign = TextAlign.Start ,
-            fontSize = fontSize.sp
+            style = MaterialTheme.typography.titleMedium
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DatePicker( modifier: Modifier = Modifier , label : String ){
+    val mContext = LocalContext.current
+    val calendar = Calendar.getInstance()
+    val year = calendar.get(Calendar.YEAR)
+    val month = calendar.get(Calendar.MONTH)
+    val day = calendar.get(Calendar.DAY_OF_MONTH)
+    calendar.time = Date()
+    val date = remember { mutableStateOf( "$day/${month+1}/$year" ) }
+    val datePickerDialog = DatePickerDialog(
+        mContext, { _: DatePicker, mYear: Int, mMonth: Int, mDayOfMonth: Int ->
+            date.value = "$mDayOfMonth/${mMonth+1}/$mYear"
+        }, year, month, day
+    )
+    Row( modifier = modifier ) {
+        TextField(
+            value = date.value,
+            onValueChange = {},
+            readOnly = true,
+            textStyle = TextStyle.Default.copy( fontSize = 20.sp ) ,
+            colors = TextFieldDefaults.textFieldColors(
+                focusedLabelColor = MaterialTheme.colorScheme.primary,
+                unfocusedLabelColor = MaterialTheme.colorScheme.primary,
+                textColor = Color.Black ,
+                containerColor = MaterialTheme.colorScheme.secondaryContainer ,
+                focusedIndicatorColor = Color.Transparent,
+                unfocusedIndicatorColor = Color.Transparent,
+                disabledIndicatorColor = Color.Transparent
+            ),
+            label = { Text(text = label) },
+            enabled = false,
+            modifier = Modifier
+                .weight(1.0f)
+                .padding(16.dp)
+                .fillMaxWidth()
+                .background(MaterialTheme.colorScheme.secondaryContainer , RoundedCornerShape(16.dp))
+                .clickable {
+                    datePickerDialog.show()
+                }
         )
     }
 }