+++ app/src/main/java/com/pict/pbl/medswift/screens/user_auth/RegisterScreen.kt
@@ -1,12 +1,10 @@
-package com.pict.pbl.medswift.screens
+package com.pict.pbl.medswift.screens.user_auth
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
@@ -15,24 +13,17 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.pict.pbl.medswift.auth.CurrentUserDetails
 import com.pict.pbl.medswift.data.User
-import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
-import com.pict.pbl.medswift.ui.theme.ScreenTitle
-import com.pict.pbl.medswift.ui.theme.TextInput
+import com.pict.pbl.medswift.ui.theme.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -43,6 +34,15 @@ class RegisterScreen : ComponentActivity() {
     private val currentUser = User()
     private val auth = Firebase.auth
     private var userPassword = ""
+    private val genderOptions = listOf( "Male" , "Female" )
+    private val bloodGroupOptions = listOf(
+        "B +ve" ,
+        "B -ve" ,
+        "O +ve" ,
+        "O -ve" ,
+        "AB +ve" ,
+        "AB -ve"
+    )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -69,6 +69,14 @@ class RegisterScreen : ComponentActivity() {
         Column( modifier = Modifier.verticalScroll( rememberScrollState() ) ){
             ScreenTitle(title = "Register", icon = Icons.Default.PersonAdd)
             TextInput(
+                label = "Email Address",
+                onValueChange = { currentUser.email = it } ,
+                keyboardType = KeyboardType.Email ,
+                icon = Icons.Default.Email ,
+                modifier = modifier
+            )
+            UserPassword(modifier = modifier)
+            TextInput(
                 label = "First Name",
                 onValueChange = { currentUser.firstName = it } ,
                 keyboardType = KeyboardType.Text ,
@@ -87,22 +95,28 @@ class RegisterScreen : ComponentActivity() {
                 icon = Icons.Default.Phone ,
                 modifier = modifier
             )
-            TextInput(
-                label = "Email Address",
-                onValueChange = { currentUser.email = it } ,
-                keyboardType = KeyboardType.Email ,
-                icon = Icons.Default.Email ,
-                modifier = modifier
+            DropDownInput(
+                items = genderOptions,
+                onItemClick = { index -> currentUser.gender = genderOptions[ index ] } ,
+                modifier = modifier,
+                label = "Gender"
             )
-            UserPassword(modifier = modifier)
-            UserGender( modifier )
-            UserBloodGroup( modifier )
+            DropDownInput(
+                items = bloodGroupOptions,
+                onItemClick = { index -> currentUser.bldGrp = bloodGroupOptions[ index ] } ,
+                modifier = modifier,
+                label = "Blood Group"
+            )
             TextInput(
                 label = "Weight",
                 onValueChange = { currentUser.weight = if( it.isNotEmpty() ) { it.toInt() } else { 0  } } ,
                 keyboardType = KeyboardType.Number ,
                 icon = Icons.Default.Accessibility ,
                 modifier = modifier
+            )
+            DatePicker(
+                modifier = modifier ,
+                label = "Select Date of Birth"
             )
             TextInput(
                 label = "Height",
@@ -129,33 +143,6 @@ class RegisterScreen : ComponentActivity() {
 
     @OptIn(ExperimentalMaterial3Api::class)
     @Composable
-    private fun UserGender(
-        modifier: Modifier = Modifier
-    ){
-        var gender by remember { mutableStateOf("") }
-        val focusManager = LocalFocusManager.current
-        TextField(
-            modifier = modifier,
-            value = gender,
-            onValueChange = {
-                    it ->
-                currentUser.gender = it
-                gender = it
-            },
-            placeholder = { Text(text = "Gender")},
-            colors = TextFieldDefaults.textFieldColors(
-                focusedLabelColor = MaterialTheme.colorScheme.primary,
-                unfocusedLabelColor = MaterialTheme.colorScheme.primaryContainer,
-                textColor = MaterialTheme.colorScheme.onPrimaryContainer
-            ),
-            singleLine = true,
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Text, imeAction = ImeAction.Next),
-            keyboardActions = KeyboardActions(onNext = { focusManager.moveFocus(FocusDirection.Next)})
-        )
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
     private fun UserPassword( modifier : Modifier ){
         var pass by rememberSaveable{ mutableStateOf( "" ) }
         var passwordVisible by rememberSaveable { mutableStateOf(false) }
@@ -167,7 +154,7 @@ class RegisterScreen : ComponentActivity() {
                 pass = it
                 userPassword = pass
             } ,
-            placeholder = { Text("Password") } ,
+            placeholder = { Text("Password",style = MaterialTheme.typography.labelSmall) } ,
             leadingIcon = { Icon( imageVector = Icons.Default.Lock , contentDescription = "Email Address" ) } ,
             visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
             trailingIcon = {
@@ -180,33 +167,6 @@ class RegisterScreen : ComponentActivity() {
                 }
             },
             keyboardOptions = KeyboardOptions( keyboardType = KeyboardType.Password , imeAction = ImeAction.Done),
-        )
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    private fun UserBloodGroup(
-        modifier: Modifier = Modifier
-    ){
-        var bloodGroup by remember { mutableStateOf("") }
-        val focusManager = LocalFocusManager.current
-        TextField(
-            modifier = modifier,
-            value = bloodGroup,
-            onValueChange = {
-                    it ->
-                currentUser.bldGrp = it
-                bloodGroup = it
-            },
-            placeholder = { Text(text = "Blood Group")},
-            colors = TextFieldDefaults.textFieldColors(
-                focusedLabelColor = MaterialTheme.colorScheme.primary,
-                unfocusedLabelColor = MaterialTheme.colorScheme.primaryContainer,
-                textColor = MaterialTheme.colorScheme.onPrimaryContainer
-            ),
-            singleLine = true,
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Text, imeAction = ImeAction.Next),
-            keyboardActions = KeyboardActions(onNext = { focusManager.moveFocus(FocusDirection.Next)})
         )
     }