+++ app/src/main/java/com/pict/pbl/medswift/ui/theme/Type.kt
@@ -2,33 +2,48 @@ package com.pict.pbl.medswift.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import com.pict.pbl.medswift.R
+
+val fonts = FontFamily(
+    Font( R.font.worksans_regular ) ,
+    Font( R.font.worksans_bold , FontWeight.Bold ) ,
+    Font( R.font.worksans_light , FontWeight.Light ) ,
+    Font( R.font.worksans_medium , FontWeight.Medium ) ,
+    Font( R.font.worksans_semibold , FontWeight.SemiBold )
+)
 
 // Set of Material typography styles to start with
+// Refer -> https://m3.material.io/styles/typography/applying-type
 val Typography = Typography(
-    bodyLarge = TextStyle(
-        fontFamily = FontFamily.Default,
+    displayLarge = TextStyle(
+        fontFamily = fonts,
+        fontWeight = FontWeight.Bold,
+        fontSize = 24.sp,
+        letterSpacing = 0.5.sp
+    ) ,
+    titleLarge = TextStyle(
+        fontFamily = fonts,
+        fontWeight = FontWeight.Bold,
+        fontSize = 22.sp,
+        letterSpacing = 0.25.sp
+    ),
+    labelLarge = TextStyle(
+        fontFamily = fonts,
+        fontWeight = FontWeight.Normal,
+        fontSize = 18.sp,
+    ),
+    labelMedium = TextStyle(
+        fontFamily = fonts,
         fontWeight = FontWeight.Normal,
         fontSize = 16.sp,
-        lineHeight = 24.sp,
-        letterSpacing = 0.5.sp
-    )
-    /* Other default text styles to override
-    titleLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 22.sp,
-        lineHeight = 28.sp,
-        letterSpacing = 0.sp
     ),
     labelSmall = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Medium,
-        fontSize = 11.sp,
-        lineHeight = 16.sp,
-        letterSpacing = 0.5.sp
-    )
-    */
+        fontFamily = fonts,
+        fontWeight = FontWeight.Normal,
+        fontSize = 14.sp,
+    ),
 )